MiPrimerProyecto/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
distlib==0.3.7
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
filelock==3.12.4
platformdirs==3.10.0
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.7.1
tzdata==2023.3
virtualenv==20.24.5